issue #123 : Move utilities to a separate package for reuse
